requirements.txt
pip
emoji==0.5.4
discord.py==1.2.5